ai-agent/requirements.txt
# AI Agent Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI & NLP
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
nltk==3.8.1
spacy==3.7.2
transformers==4.36.0

# WhatsApp & Communication  
requests==2.31.0
twilio==8.11.0
webhooks==0.6.0

# Google Calendar
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
supabase==2.0.2

# Data Processing
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2
pytz==2023.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic-settings==2.1.0
aiofiles==23.2.1
jinja2==3.1.2